Remove color-scheme-specific icon
Replace color-scheme-specific icon svg with a general one
with hex color #AEAEAE (obtained from the quadratic mean of the
quadratic means of each of chrome's version 120.0.6099.71 color
scheme default grey foreground color and background color in version).

To-do: replace the backup ico and png versions of the icon because
only the svg was updated.